WIP
 - added `native_iam_role` service
 - changed name of the `native_iam_oauth` to `native_iam_auth`
 - removed TS tests
 - added golang tests
 - added new `docker-compose.all-in-one.dev.yml` wich forwards ports to the `127.0.0.1` and makes it easier to develop applications
 - a lot of other changes :)

+++ modules/native/services/namespace/src/services/namespace.go
@@ -101,7 +101,7 @@ type NamespaceInMongo struct {
 
 	Updated time.Time `bson:"updated"`
 	Created time.Time `bson:"created"`
-	Version uint32    `bson:"version"`
+	Version uint64    `bson:"version"`
 }
 
 func (n *NamespaceInMongo) ToGRPCNamespace() *grpc.Namespace {
